+++ trl/experimental/distillation/distillation_trainer.py
@@ -103,19 +103,6 @@ def _print_completions_sample(prompts: list[str], completions: list[str], step: 
     console.print(panel)
 
 
-def _add_tail_bucket(log_probs, valid_mask):
-    """Append a (K+1)-th tail element: log(1 - sum(exp(top_k_logps))).
-
-    This creates a proper probability distribution over K+1 elements, preventing trivial zero loss when top_k is small
-    (especially top_k=1).
-    """
-    log_sum = torch.logsumexp(log_probs, dim=-1, keepdim=True)
-    log_sum = torch.clamp(log_sum, max=-1e-7)  # ensure sum < 1
-    tail = torch.log(-torch.expm1(log_sum))  # log(1 - exp(log_sum))
-    tail_mask = torch.ones_like(valid_mask[..., :1], dtype=torch.bool)
-    return torch.cat([log_probs, tail], dim=-1), torch.cat([valid_mask, tail_mask], dim=-1)
-
-
 def _jsd_divergence(student_log_probs, teacher_log_probs, beta, support_mask=None):
     """Compute JSD (or forward/reverse KL) from log-probability tensors.
 
@@ -540,8 +527,6 @@ class DistillationTrainer(_BaseTrainer):
         self.temperature = args.temperature
         self.top_p = args.top_p
         self.num_generations = args.num_generations
-        self.loss_top_k = args.loss_top_k
-        self.loss_add_tail = args.loss_add_tail
 
         # ── Buffer state ──
         self._buffered_inputs = None
@@ -1041,12 +1026,10 @@ class DistillationTrainer(_BaseTrainer):
         beta=0.5,
         temperature=1.0,
         reduction="batchmean",
-        top_k=0,
-        add_tail=True,
         num_items_in_batch=None,
     ):
         """
-        Compute the generalized Jensen-Shannon Divergence loss for knowledge distillation.
+        Compute the generalized Jensen-Shannon Divergence loss for knowledge distillation over the full vocabulary.
 
         Args:
             student_logits: Tensor of shape (batch_size, sequence_length, vocab_size).
@@ -1055,12 +1038,6 @@ class DistillationTrainer(_BaseTrainer):
             beta: Interpolation coefficient. 0.0 = forward KL, 1.0 = reverse KL.
             temperature: Softmax temperature.
             reduction: 'batchmean', 'sum', 'mean', or 'none'.
-            top_k: Number of top tokens to restrict the loss to. The support set depends on beta:
-                beta=0 (forward KL) uses teacher's top-k, beta=1 (reverse KL) uses student's top-k, 0<beta<1 (JSD) uses
-                the union of both. Distributions are re-normalized over the selected support. If 0, the full vocabulary
-                is used.
-            add_tail: Whether to append a tail bucket representing the remaining probability mass
-                outside the selected top-k support.
 
         Returns:
             Scalar loss tensor.
@@ -1068,55 +1045,10 @@ class DistillationTrainer(_BaseTrainer):
         student_logits = student_logits / temperature
         teacher_logits = teacher_logits / temperature
 
-        support_mask = None
+        student_log_probs = F.log_softmax(student_logits, dim=-1)
+        teacher_log_probs = F.log_softmax(teacher_logits, dim=-1)
 
-        if top_k > 0 and student_logits.size(-1) > top_k:
-            neg_inf = torch.full((), float("-inf"), dtype=student_logits.dtype, device=student_logits.device)
-            student_log_probs_full = F.log_softmax(student_logits, dim=-1)
-            teacher_log_probs_full = F.log_softmax(teacher_logits, dim=-1)
-
-            if beta == 0:
-                # Forward KL: teacher-selected support
-                _, support = teacher_logits.topk(top_k, dim=-1)
-                support_mask = torch.ones_like(support, dtype=torch.bool)
-            elif beta == 1:
-                # Reverse KL: student-selected support
-                _, support = student_logits.topk(top_k, dim=-1)
-                support_mask = torch.ones_like(support, dtype=torch.bool)
-            else:
-                # JSD: union of both supports (concatenate + deduplicate)
-                _, student_top = student_logits.topk(top_k, dim=-1)
-                _, teacher_top = teacher_logits.topk(top_k, dim=-1)
-                support = torch.cat([teacher_top, student_top], dim=-1)
-                support_mask = torch.ones(support.shape, dtype=torch.bool, device=support.device)
-                for i in range(1, support.shape[-1]):
-                    prev_matches = support[..., i : i + 1] == support[..., :i]
-                    prev_valid = support_mask[..., :i]
-                    support_mask[..., i] &= ~(prev_matches & prev_valid).any(dim=-1)
-                support = torch.where(support_mask, support, torch.zeros_like(support))
-
-            student_support_logps = student_log_probs_full.gather(-1, support)
-            teacher_support_logps = teacher_log_probs_full.gather(-1, support)
-
-            # Mask invalid (duplicate) positions with -inf
-            student_topk_logps = torch.where(support_mask, student_support_logps, neg_inf)
-            teacher_topk_logps = torch.where(support_mask, teacher_support_logps, neg_inf)
-
-            if add_tail:
-                # Add tail bucket: append log(1 - sum(exp(top_k_logps))) to preserve
-                # the remaining probability mass outside the top-k. This prevents trivial
-                # zero loss when top_k is small (especially top_k=1).
-                base_support_mask = support_mask
-                student_log_probs, support_mask = _add_tail_bucket(student_topk_logps, base_support_mask)
-                teacher_log_probs, _ = _add_tail_bucket(teacher_topk_logps, base_support_mask)
-            else:
-                student_log_probs = student_topk_logps - torch.logsumexp(student_topk_logps, dim=-1, keepdim=True)
-                teacher_log_probs = teacher_topk_logps - torch.logsumexp(teacher_topk_logps, dim=-1, keepdim=True)
-        else:
-            student_log_probs = F.log_softmax(student_logits, dim=-1)
-            teacher_log_probs = F.log_softmax(teacher_logits, dim=-1)
-
-        jsd = _jsd_divergence(student_log_probs, teacher_log_probs, beta, support_mask)
+        jsd = _jsd_divergence(student_log_probs, teacher_log_probs, beta)
         return DistillationTrainer._reduce_divergence_loss(
             jsd, labels=labels, reduction=reduction, num_items_in_batch=num_items_in_batch
         )
@@ -1157,8 +1089,6 @@ class DistillationTrainer(_BaseTrainer):
             labels=labels,
             beta=self.beta,
             temperature=self.temperature,
-            top_k=self.loss_top_k,
-            add_tail=self.loss_add_tail,
             num_items_in_batch=num_items_in_batch,
         )
